requirements.txt
streamlit
google-genai
google-search-results
python-dotenv
requests
